% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ffars.R
\name{fars_map_state}
\alias{fars_map_state}
\title{Map accidents by state and year}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{The numerical ID of a US state}

\item{year}{the year}
}
\value{
A map of the state and location of accident - an image
}
\description{
Plots the location of accidents on maps of US states
}
\note{
Requires dplyr and maps package. To be called at package level via 'importFrom'
statements

Errors on state numbers not included in the data package
}
\examples{
fars_map_state(48, 2013)
}

