% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ffars.R
\name{fars_read}
\alias{fars_read}
\title{Read Fatality Analysis Reporting System (FARS) data
Read FARS data in comma-separated files (compression optional) into \code{tbl_df}
Fails if filename not available - use \link{make_filename}  for canonical file nameing}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{- the filename, a bzipped .csv file supplied with the assignment}
}
\value{
The data as a \code{tbl_df}
}
\description{
Read Fatality Analysis Reporting System (FARS) data
Read FARS data in comma-separated files (compression optional) into \code{tbl_df}
Fails if filename not available - use \link{make_filename}  for canonical file nameing
}
\examples{
\dontrun{fars_read("accident_2013.csv.bz2")}
}

