% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ffars.R
\name{fars_read_years}
\alias{fars_read_years}
\title{Read years}
\usage{
fars_read_years(years)
}
\value{
A \code{tbl_df} with month and years as columns
}
\description{
Read FARS data into memory and select month and years
}
\note{
Uses \code{\link{make_filename}} internally, warns on invalid year
}
\examples{
fars_read_years(c(2013, 2014))
}

