% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ffars.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{Summarize years by month}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{vector of years}
}
\value{
tbl_df with MONTH and the respective years as columns
}
\description{
Summarize years by month
}
\examples{
fars_summarize_years(c(2013, 2014))
}

