% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ffars.R
\name{make_filename}
\alias{make_filename}
\title{Create canonical filenmae}
\usage{
make_filename(year)
}
\arguments{
\item{year}{- the year as four digits. Data set includes only 2013-2015}
}
\value{
the filename styled including the ending for bz2-compressed csv
}
\description{
Create canonical filenmae
}
\note{
Internal function, will blindly assume names without any checking of existence of file
}
\examples{
make_filename(2013)
}

